flaskapp/requirements.txt
Flask
requests
python-dotenv
SQLAlchemy
Flask_SQLAlchemy
gunicorn
authlib
pymysql
pycryptodome
